[DPP R2] Added support for DPP R2 events
Added support for DPP R2 events that provide additional
details about the onboarding process of a remote enrollee.
Specifically, DPP R2 configurator waits for response from
the enrollee, which reports back the SSID, tried channels
and band support in case it cannot find the AP. When it
reports success, then it means that it is acutally connected.

Bug: 139381558
Test: Manual tests with DPP R1 and R2 enrollees
Test: atest DppManagerTest
Test: atest WifiManagerTest
Test: atest EasyConnectStatusCallbackTest
Test: act.py -c ../WifiDppConfig.json -tc WifiDppTest
Change-Id: I55a8b1e25737e90e304f1933d09ffcd6bcae6ef9