Enforce that registered authenticators are not null
OEMs should ensure that the config_biometric_sensors
overlay for core/res/res/values.xml is correct. The majority
of devices already have fingerprint sensors, thus to ease
migration, the default configuration is for a "strong
fingerprint sensor on ID0". On devices where this is not true,
the config must be overlaid with whatever sensor the device
actually has.

Bug: 141025588

Test: atest BiometricServiceTest
Change-Id: I61f56d4f11fb7ce3d164a3528a8ecc3522c7e342