Move the app to /system_ext
We plan to forbid hidden API in product partition.
So these apps have to be moved to /system_ext because it uses hidden API.
It is not permanent, so they can be moved back to /product due to
needs if there isn't hidden usage anymore.

Bug: 137908189
Test: m
Test: check whether device boots successfully
Change-Id: I0662674a66bc54ee4713e5c7ed6ad0db4d342338
Merged-In: I0662674a66bc54ee4713e5c7ed6ad0db4d342338