Revert "Use framework-appsearch-stubs instead of framework-appsearch"
This reverts commit 858946274e5ff28bcbc7980b3522c74f539d44be.

Reason for revert: There are ~20 pending CLs in this project in various
stages of review, and this CL causes widespread breakages in many parts
of our work (including CTS tests, coretests and servicetests) and blocks
multiple team members. Given the closeness of the API review deadline
and the fact that framework-appsearch is empty at HEAD (so building at
head is not indicative that the infrastructure supports the project's
needs), would it be possible to hold off on building against the stubs
or now until the various tests are set up at HEAD? ETA for getting code
submitted in this project: 1 day.

Change-Id: Ie2ea82f5ee41861e93f5bdf34ceae1417328d26a