On reboot session handling.
- re-seal regular sessions,
- re-seal+validate APEX sessions.

Previously we were validating them as well. This is unnecessary as
validation will happen on commit/transfer calls later.

This was introduced in 68eaddd8d7cd83b6eb3c8f24a19481b600766c9a when we
were using mSealed flag for both sealing and validation. It's not the
case anymore and we can safely deffer validation to commit/transfer.

Test: atest PackageManagerShellCommandTest SplitTests StagedRollbackTest CtsAtomicInstallTestCases PackageInstallerSessionTest InstallSessionTransferTest
Bug: b/136132412

Change-Id: I1ccb29ad41e25561f4f54ea51da3b9a725dd5647