Kill app when OP_MANAGE_EXTERNAL_STORAGE is denied.
When an app requests the special access permission to be able to manage
external storage, and it is denied, we need to kill the app, in order to
lose the external_storage gid it may have that allows it access to
unreliable USB OTG volumes.

When the app op is granted, we don't kill the app to pickup the new
gid because it may result in bad UX especially since that gid is only
useful to access unreliable USB OTGs that are rarely present.
The next time the app is restarted, it will get the external_storage
gid and will be able to access any mounted USB OTG volumes.

Bug: 144914977
Test: Copy paste of I73c5ce56ee24e5c864d7c9280b83fc3854701a97
Change-Id: I5cb13dc9c51cb8b4e942e529b4a1c7dac180cace