Fix View backward navigation bug
b/147829061 used View id to search for the next backward view. The
problem was that there might be several Views with the same ID in View
hierarchy, so it may return the wrong View with the same ID. To fix
that, ag/10111732 compared the View itself. The idea was right but there
was a bug in that CL. This CL fixed the original bug and the following
bugs.

Fixes: 148545035
Fixes: 148166620
Bug: 147829061
Test: atest TimePickerTest
Test: atest BackwardNavigationTest
Test: atest SpinnerTest
Change-Id: I4ec44f91bbaa86b0f52e10611b91fcceb2588d12