If an app cancels an experimental bubble, update the dot to hide
We don't honour app cancels for experimental bubbles, but we should
hide the dot on the bubble when the app cancels a notification for
an experimental bubble. Typically if an app cancels a notification
it means that content is "read" (at least for bubbles) so the dot
shouldn't be necessary.

Test: - get an experimental bubble from some conversation
      - navigate to that conversation in the app, not the bubble
      => Note that the notification is gone from the shade and the
         bubble no longer has a dot
Fixes: 147232318

Change-Id: If2a89896c7ee0c2264385c6fb501569abdc02b22